# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "erdsea-api",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5000",
    "paths": {
...
        "/collections/create": {
            "post": {
                "description": "Creates a collection with given info.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Creates a collection.",
                "parameters": [
                    {
                        "description": "collection info",
                        "name": "createCollectionRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.CreateCollectionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.Collection"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    },
                    "505": {
                        "description": "HTTP Version Not Supported",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    }
                }
            }
        },
        "/collections/list/{offset}/{limit}": {
            "get": {
                "description": "Retrieves a list of collections. Unsorted.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Gets collections.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.Collection"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    }
                }
            }
        },
        "/collections/{collectionName}": {
            "get": {
                "description": "Retrieves a collection by its name.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Gets collection by name.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "collection name",
                        "name": "collectionName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.Collection"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Sets info for a collection.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Set collection info.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "collection name",
                        "name": "collectionName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "collection info",
                        "name": "updateCollectionRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.UpdateCollectionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.Collection"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    }
                }
            }
        },
        "/collections/{collectionName}/assets/{offset}/{limit}": {
            "get": {
                "description": "Retrieves the assets of a collection. Unsorted.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Get collection assets.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "collection name",
                        "name": "collectionName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.Asset"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    }
                }
            }
        },
        "/collections/{collectionName}/cover": {
            "get": {
                "description": "Retrieves a collection cover image. It will be sent as base64 encoding (sdt, raw) of its byte representation.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Get collection cover image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "collection name",
                        "name": "collectionName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Expects base64 encoding (sdt, raw) of the image representation. Returns empty string. Max size of byte array is 1MB.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Set collection cover image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "collection name",
                        "name": "collectionName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "base64 encoded image",
                        "name": "image",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "primitive"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    }
                }
            }
        },
        "/collections/{collectionName}/profile": {
            "get": {
                "description": "Retrieves a collection cover image. It will be sent as base64 encoding (sdt, raw) of its byte representation.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Get collection profile image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "collection name",
                        "name": "collectionName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Expects base64 encoding (sdt, raw) of the image representation. Returns empty string. Max size of byte array is 1MB.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Set collection profile image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "collection name",
                        "name": "collectionName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "base64 encoded image",
                        "name": "image",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "primitive"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    }
                }
            }
        },
        "/collections/{collectionName}/statistics": {
            "post": {
                "description": "Gets statistics for a collection. It will be cached for 10 minutes.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Gets collection statistics.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "collection name",
                        "name": "collectionName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.CollectionStatistics"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    }
                }
            }
        },
...
        "/search/collections/{collectionName}": {
            "get": {
                "description": "Searches for collections by name. Cached for 20 minutes. Limit 5 elements.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Search collections by name.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search string",
                        "name": "collectionName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.Collection"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "data.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                }
            }
        },
        "data.Asset": {
            "type": "object",
            "properties": {
                "collectionID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "link": {
                    "type": "string"
                },
                "listed": {
                    "type": "boolean"
                },
                "nonce": {
                    "type": "integer"
                },
                "ownerId": {
                    "type": "integer"
                },
                "priceNominal": {
                    "type": "number"
                },
                "royaltiesPercent": {
                    "type": "number"
                },
                "tokenID": {
                    "type": "string"
                }
            }
        },
        "data.Collection": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "integer"
                },
                "creatorID": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "discordLink": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "instagramLink": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "telegramLink": {
                    "type": "string"
                },
                "tokenID": {
                    "type": "string"
                },
                "twitterLink": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
...
        "services.CreateCollectionRequest": {
            "type": "object",
            "properties": {
                "collectionName": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "discordLink": {
                    "type": "string"
                },
                "instagramLink": {
                    "type": "string"
                },
                "telegramLink": {
                    "type": "string"
                },
                "tokenId": {
                    "type": "string"
                },
                "twitterLink": {
                    "type": "string"
                },
                "userAddress": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
...
        "services.UpdateCollectionRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "discordLink": {
                    "type": "string"
                },
                "instagramLink": {
                    "type": "string"
                },
                "telegramLink": {
                    "type": "string"
                },
                "twitterLink": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        }
    }
}